--- app/mypage/_lib/getUserMe.ts
@@ -1,13 +1,0 @@
-import customFetch from "@/services/customFetch";
-
-export default async function getUserMe() {
-  const response = await customFetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/v1/users/me`,
-  );
-
-  if (!response.ok) {
-    throw new Error(response.statusText);
-  }
-
-  return response.json();
-}
